+++ src/plotting.jl
@@ -1,4 +1,23 @@
 ##########################################################################################
+# Attractors
+##########################################################################################
+"""
+    plot_attractors(attractors::Dict{Int, StateSpaceSet}; kwargs...)
+
+Plot the attractors as a scatter plot.
+
+## Keyword arguments
+
+- All the [common plotting keywords](@ref common_plot_kwargs).
+  Particularly important is the `access` keyword.
+- `sckwargs = (strokewidth = 0.5, strokecolor = :black,)`: additional keywords
+  propagated to the `Makie.scatter` function that plots the attractors.
+"""
+function plot_attractors end
+function plot_attractors! end
+export plot_attractors, plot_attractors!
+
+##########################################################################################
 # Basins
 ##########################################################################################
 """